moved header to container and updtaed filter components
