+++ src/App.js
@@ -1,6 +1,6 @@
 import { BrowserRouter } from 'react-router-dom';
 import Routes from './config/Routes';
-import Header from './components/Header';
+import Header from './containers/Header';
 import './styles/App.css';
 
 function App() {
